requirements.txt
# for download data initially
# datalad # note: pip-installing didn't work, had to do sudo apt-get install datalad

# actual reqs
pathlib
h5py
tables
numpy
pandas
torch
transformers
datasets
scikit-learn
imodels
skorch
nltk
fire
dict_hash
# imodelsx should be installed from source
# improved parallelization
# vllm
# ray
# pydantic==1.10.15

# basic
tqdm
# ridge_utils